+++ carrito/carrito/src/main/java/com/example/carrito/dto/ProductoResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 public class ProductoResponse {
     private Long id;
     private String nombre;
-    private Double precio;
+    private int precio;
 
 }
